PI SaaS Actions: various fixes
* Error codes are parsed as longs, not ints.
* Fix assertEquals parameter order to match (expected, actual).

+++ src/main/java/com/veritomyx/actions/DeleteAction.java
@@ -46,7 +46,7 @@ public class DeleteAction extends BaseAction {
 	}
 
 	@Override
-	public int getErrorCode() {
+	public long getErrorCode() {
 		preCheck();
 		return super.getErrorCode();
 	}
