+++ src/main/java/com/veritomyx/actions/PrepAction.java
@@ -65,7 +65,7 @@ public class PrepAction extends BaseAction {
 	}
 
 	@Override
-	public int getErrorCode() {
+	public long getErrorCode() {
 		preCheck();
 		return super.getErrorCode();
 	}
