+++ src/main/java/com/veritomyx/actions/RunAction.java
@@ -63,7 +63,7 @@ public class RunAction extends BaseAction {
 	}
 
 	@Override
-	public int getErrorCode() {
+	public long getErrorCode() {
 		preCheck();
 		return super.getErrorCode();
 	}
