+++ src/main/java/com/veritomyx/actions/StatusAction.java
@@ -76,7 +76,7 @@ public class StatusAction extends BaseAction {
 	}
 
 	@Override
-	public int getErrorCode() {
+	public long getErrorCode() {
 		preCheck();
 		return super.getErrorCode();
 	}
